Split code and make it work again
