+++ 3696/2/commands/start.rb
@@ -7,9 +7,6 @@ require_relative '../helpers/base_command_helpers'
 module StartCommand
   include BaseCommandHelpers
 
-  DATA_PATH = './data/numbers.yaml'
-  NUMBERS_LIST_KEY = 'numbers'
-
   def start!(*)
     if already_registered?
       stop_message
@@ -28,6 +25,13 @@ module StartCommand
   end
 
   private
+
+  DATA_PATH = './data/numbers.yaml'
+  NUMBERS_LIST_KEY = 'numbers'
+
+  def numbers
+    @numbers ||= YAML.load_file(DATA_PATH).fetch(NUMBERS_LIST_KEY, []).map(&:to_i)
+  end
 
   def write_session_register(number = nil, *)
     reversed_redis = Redis.new
@@ -49,10 +53,6 @@ module StartCommand
   def register_user(redis, number)
     session[:number] = number
     redis.set(number, user_id)
-  end
-
-  def numbers
-    @numbers ||= YAML.load_file(DATA_PATH).fetch(NUMBERS_LIST_KEY, []).map(&:to_i)
   end
 
   def rescue_number
